rfac: refactor app drawer
